#42- 리덕스 분리  적용
#42- 리덕스 분리  적용
